refactor: 이미지 등록 커스텀 hook 구현
#34
